Struct arithmetic_eval::arith::StdArithmetic

pub struct StdArithmetic;

Arithmetic on a number type that implements all necessary operations natively.

As an example, this type implements Arithmetic for f32, f64, and the floating-point complex numbers from the num-complex crate.