Enum arithmetic_eval::Value

#[non_exhaustive]
pub enum Value<T> {
    Prim(T),
    Bool(bool),
    Function(Function<T>),
    Tuple(Tuple<T>),
    Object(Object<T>),
    Ref(OpaqueRef),
}

Values produced by expressions during their interpretation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Prim(T)

Primitive value, such as a number. This does not include Boolean values, which are a separate variant.

Literals must necessarily map to primitive values, but there may be some primitive values not representable as literals.

Bool(bool)

Boolean value.

Function(Function<T>)

Function.

Tuple(Tuple<T>)

Tuple of zero or more values.

Object(Object<T>)

Object with zero or more named fields.

Ref(OpaqueRef)

Opaque reference to a native value.

Implementations

impl<T> Value<T>

pub fn native_fn(function: impl NativeFn<T> + 'static) -> Self

Creates a value for a native function.

pub fn wrapped_fn<const CTX: bool, Args, F>(fn_to_wrap: F) -> Selfwhere FnWrapper<Args, F, CTX>: NativeFn<T> + 'static,

Creates a wrapped function.

Calling this method is equivalent to wrapping a function and calling Self::native_fn() on it. Thanks to type inference magic, the Rust compiler will usually be able to extract the Args type param from the function definition, provided that type of function arguments and its return type are defined explicitly or can be unequivocally inferred from the declaration.

pub const fn void() -> Self

Creates a void value (an empty tuple).

pub fn opaque_ref(value: impl Any + Debug + PartialEq) -> Self

Creates a reference to a native variable.

pub fn value_type(&self) -> ValueType

Returns the type of this value.

pub fn is_void(&self) -> bool

Checks if this value is void (an empty tuple).

pub fn is_function(&self) -> bool

Checks if this value is a function.