Module arithmetic_eval::error

Evaluation errors.

Structs

BacktraceElement

Element of a backtrace, i.e., a function / method call.

Error

Evaluation error together with one or more relevant code spans.

ErrorWithBacktrace

Error with the associated backtrace.

LocationInModule

Code fragment together with information about the module containing the fragment.

Enums

ArithmeticError

Arithmetic errors raised by Arithmetic operations on primitive values.

AuxErrorInfo

Auxiliary information about an evaluation error.

ErrorKind

Kinds of errors that can occur when compiling or interpreting expressions and statements.

RepeatedAssignmentContext

Context for ErrorKind::RepeatedAssignment.

TupleLenMismatchContext

Context for ErrorKind::TupleLenMismatch.

UnsupportedType

Description of a construct not supported by a certain module (e.g., interpreter or type inference).

Type Aliases

EvalResult

Result of an expression evaluation.