Function arithmetic_eval::fns::wrap

pub const fn wrap<const CTX: bool, T, F>(function: F) -> FnWrapper<T, F, CTX>

Wraps a function enriching it with the information about its arguments. This is a slightly shorter way to create wrappers compared to calling FnWrapper::new().

See FnWrapper for more details on function requirements.