Type Alias arithmetic_eval::fns::Unary

pub type Unary<T> = FnWrapper<(T, T), fn(_: T) -> T>;

Unary function wrapper.

Aliased Type

struct Unary<T> { /* private fields */ }

Implementations

impl<T, F, const CTX: bool> FnWrapper<T, F, CTX>

pub const fn new(function: F) -> Self

Creates a new wrapper.

Note that the created wrapper is not guaranteed to be usable as NativeFn. For this to be the case, function needs to be a function or an Fn closure, and the T type argument needs to be a tuple with the function return type and the argument types (in this order).