Trait arithmetic_eval::NativeFn

pub trait NativeFn<T> {
    fn evaluate<'a>(
        &self, 
        args: Vec<SpannedValue<'a, T>>, 
        context: &mut CallContext<'_, 'a, T>
    ) -> EvalResult<'a, T>;
}

Function on zero or more Values.

Native functions are defined in the Rust code and then can be used from the interpreted code. See fns module docs for different ways to define native functions.

Required methods

fn evaluate<'a>(
&self,
args: Vec<SpannedValue<'a, T>>,
context: &mut CallContext<'_, 'a, T>
) -> EvalResult<'a, T>

Executes the function on the specified arguments.

Loading content...