Trait arithmetic_eval::Number

pub trait Number: Clone + 'static { }

Marker trait for possible literals.

This trait is somewhat of a crutch, necessary to ensure that function wrappers can accept number arguments and distinguish them from other types (booleans, vectors, tuples, etc.).