Trait arithmetic_eval::VariableMap

pub trait VariableMap<'a, T> {
    fn get_variable(&self, name: &str) -> Option<Value<'a, T>>;

    fn compile_module<Id, G>(
        &self, 
        id: Id, 
        block: &Block<'a, G>
    ) -> Result<ExecutableModule<'a, T>, Error<'a>>
    where
        Id: ModuleId,
        G: Grammar<'a, Lit = T>,
        T: Clone + Debug,
    { ... }
}

Encapsulates read access to named variables.

Required methods

fn get_variable(&self, name: &str) -> Option<Value<'a, T>>

Returns value of the named variable, or None if it is not defined.

Loading content...
Provided methods

fn compile_module<Id, G>(
&self,
id: Id,
block: &Block<'a, G>
) -> Result<ExecutableModule<'a, T>, Error<'a>> where
Id: ModuleId,
G: Grammar<'a, Lit = T>,
T: Clone + Debug,

Creates a module based on imports solely from this map.

The default implementation is reasonable for most cases.

Loading content...