Trait arithmetic_eval::arith::OrdArithmetic

pub trait OrdArithmetic<T>: Arithmetic<T> {
    fn partial_cmp(&self, x: &T, y: &T) -> Option<Ordering>;
}

Extends an Arithmetic with a comparison operation on values.

Required methods

fn partial_cmp(&self, x: &T, y: &T) -> Option<Ordering>

Compares two values. Returns None if the numbers are not comparable, or the comparison result otherwise.

Loading content...