arithmetic_eval::arith

Trait OrdArithmetic

pub trait OrdArithmetic<T>: Arithmetic<T> {
    // Required method
    fn partial_cmp(&self, x: &T, y: &T) -> Option<Ordering>;
}

Extends an Arithmetic with a comparison operation on values.

Required Methods

fn partial_cmp(&self, x: &T, y: &T) -> Option<Ordering>

Compares two values. Returns None if the numbers are not comparable, or the comparison result otherwise.