arithmetic_eval::env

Struct Environment

pub struct Environment<T> { /* private fields */ }

Environment containing named Values.

Note that the environment implements the Index trait, which allows to eloquently access or modify environment. Similarly, IntoIterator / Extend traits allow to construct environments.

Examples

use arithmetic_eval::{env::{Comparisons, Prelude}, Environment, Value};

// Load environment from the standard containers.
let mut env = Environment::<f64>::new();
env.extend(Prelude::iter().chain(Comparisons::iter()));
// Add a custom variable for a good measure.
env.insert("x", Value::Prim(1.0));

assert_eq!(env["true"], Value::Bool(true));
assert_eq!(env["x"], Value::Prim(1.0));
for (name, value) in &env {
    println!("{name} -> {value:?}");
}

// It's possible to base an environment on other env, as well.
let mut other_env = Environment::new();
other_env.extend(
    env.into_iter().filter(|(_, val)| val.is_function()),
);
assert!(other_env.get("x").is_none());

Implementations

impl<T> Environment<T>
where StdArithmetic: OrdArithmetic<T>,

pub fn new() -> Self

Creates a new environment.

impl<T> Environment<T>

pub fn with_arithmetic<A>(arithmetic: A) -> Self
where A: OrdArithmetic<T> + 'static,

Creates an environment with the specified arithmetic.

pub fn get(&self, name: &str) -> Option<&Value<T>>

Gets a variable by name.

pub fn contains(&self, name: &str) -> bool

Checks if this environment contains a variable with the specified name.

pub fn iter(&self) -> impl Iterator<Item = (&str, &Value<T>)> + '_

Iterates over variables.

pub fn insert(&mut self, name: &str, value: Value<T>) -> &mut Self

Inserts a variable with the specified name.

pub fn insert_native_fn( &mut self, name: &str, native_fn: impl NativeFn<T> + 'static, ) -> &mut Self

Inserts a native function with the specified name.

pub fn insert_wrapped_fn<const CTX: bool, Args, F>( &mut self, name: &str, fn_to_wrap: F, ) -> &mut Self
where FnWrapper<Args, F, CTX>: NativeFn<T> + 'static,

Inserts a wrapped function with the specified name.

Calling this method is equivalent to wrapping a function and calling insert_native_fn() on it. Thanks to type inference magic, the Rust compiler will usually be able to extract the Args type param from the function definition, provided that type of function arguments and its return type are defined explicitly or can be unequivocally inferred from the declaration.