arithmetic_eval::fns

Struct FnWrapper

pub struct FnWrapper<T, F, const CTX: bool = false> { /* private fields */ }

Wrapper of a function containing information about its arguments.

Using FnWrapper allows to define native functions with minimum boilerplate and with increased type safety. FnWrappers can be constructed explicitly or indirectly via Environment::insert_wrapped_fn(), Value::wrapped_fn(), or wrap().

Arguments of a wrapped function must implement TryFromValue trait for the applicable grammar, and the output type must implement IntoEvalResult. If you need CallContext (e.g., to call functions provided as an argument), it should be specified as a first argument.

Examples

Basic function

use arithmetic_eval::{fns, Environment, ExecutableModule, Value};

let max = fns::wrap(|x: f32, y: f32| if x > y { x } else { y });

let program = "max(1, 3) == 3 && max(-1, -3) == -1";
let program = Untyped::<F32Grammar>::parse_statements(program)?;
let module = ExecutableModule::new("test_max", &program)?;

let mut env = Environment::new();
env.insert_native_fn("max", max);
assert_eq!(module.with_env(&env)?.run()?, Value::Bool(true));

Fallible function with complex args

fn zip_arrays(xs: Vec<f32>, ys: Vec<f32>) -> Result<Vec<(f32, f32)>, String> {
    if xs.len() == ys.len() {
        Ok(xs.into_iter().zip(ys).map(|(x, y)| (x, y)).collect())
    } else {
        Err("Arrays must have the same size".to_owned())
    }
}

let program = "zip((1, 2, 3), (4, 5, 6)) == ((1, 4), (2, 5), (3, 6))";
let program = Untyped::<F32Grammar>::parse_statements(program)?;
let module = ExecutableModule::new("test_zip", &program)?;

let mut env = Environment::new();
env.insert_wrapped_fn("zip", zip_arrays);
assert_eq!(module.with_env(&env)?.run()?, Value::Bool(true));

Using CallContext to call functions

fn map_array(
    context: &mut CallContext<'_, f32>,
    array: Vec<Value<f32>>,
    map_fn: Function<f32>,
) -> Result<Vec<Value<f32>>, Error> {
    array
        .into_iter()
        .map(|value| {
            let arg = context.apply_call_location(value);
            map_fn.evaluate(vec![arg], context)
        })
        .collect()
}

let program = "map((1, 2, 3), |x| x + 3) == (4, 5, 6)";
let module = Untyped::<F32Grammar>::parse_statements(program)?;
let module = ExecutableModule::new("test", &module)?;

let mut env = Environment::new();
env.insert_wrapped_fn("map", map_array);
assert_eq!(module.with_env(&env)?.run()?, Value::Bool(true));

Implementations

impl<T, F, const CTX: bool> FnWrapper<T, F, CTX>

pub const fn new(function: F) -> Self

Creates a new wrapper.

Note that the created wrapper is not guaranteed to be usable as NativeFn. For this to be the case, function needs to be a function or an Fn closure, and the T type argument needs to be a tuple with the function return type and the argument types (in this order).