Trait arithmetic_eval::fns::TryFromValue

pub trait TryFromValue<'a, T>: Sized {
    fn try_from_value(value: Value<'a, T>) -> Result<Self, FromValueError>;
}

Fallible conversion from Value to a function argument.

This trait is implemented for base value types (such as Numbers, Functions, Values), and for two container types: vectors and tuples.

Required methods

fn try_from_value(value: Value<'a, T>) -> Result<Self, FromValueError>

Attempts to convert value to a type supported by the function.

Loading content...