Trait arithmetic_eval::fns::TryFromValue

pub trait TryFromValue<T>: Sized {
    // Required method
    fn try_from_value(value: Value<T>) -> Result<Self, FromValueError>;
}

Fallible conversion from Value to a function argument.

This trait is implemented for base value types (such as Numbers, Functions, Values), and for two container types: vectors and tuples.

Required Methods

fn try_from_value(value: Value<T>) -> Result<Self, FromValueError>

Attempts to convert value to a type supported by the function.

Object Safety

This trait is not object safe.