arithmetic_eval::fns

Trait TryFromValue

pub trait TryFromValue<T>: Sized {
    // Required method
    fn try_from_value(value: Value<T>) -> Result<Self, FromValueError>;
}

Fallible conversion from Value to a function argument.

This trait is implemented for base value types (such as Numbers, Functions, Values), and for two container types: vectors and tuples.

Required Methods

fn try_from_value(value: Value<T>) -> Result<Self, FromValueError>

Attempts to convert value to a type supported by the function.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.