arithmetic_eval

Trait NativeFn

pub trait NativeFn<T> {
    // Required method
    fn evaluate(
        &self,
        args: Vec<SpannedValue<T>>,
        context: &mut CallContext<'_, T>,
    ) -> EvalResult<T>;
}

Function on zero or more Values.

Native functions are defined in the Rust code and then can be used from the interpreted code. See fns module docs for different ways to define native functions.

Required Methods

fn evaluate( &self, args: Vec<SpannedValue<T>>, context: &mut CallContext<'_, T>, ) -> EvalResult<T>

Executes the function on the specified arguments.