arithmetic_eval

Trait Number

pub trait Number: Clone + 'static { }

Marker trait for possible literals.

This trait is somewhat of a crutch, necessary to ensure that function wrappers can accept number arguments and distinguish them from other types (booleans, vectors, tuples, etc.).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.